[update] getWinningPrice long 반환

+++ src/main/java/lotto/LottoResult.java
@@ -22,18 +22,19 @@ public class LottoResult {
     }
 
     public Float getYield() {
-        int totalWinningCount = getWinningPrice(WinningPrice.FIRST) +
+        long totalWinningCount = getWinningPrice(WinningPrice.FIRST) +
                 getWinningPrice(WinningPrice.SECOND) +
                 getWinningPrice(WinningPrice.THIRD) +
                 getWinningPrice(WinningPrice.FORTH);
         return (float) totalWinningCount / (purchasedLottos.size() * LottoPaper.LOTTO_PRICE);
     }
 
-    private Integer getWinningPrice(WinningPrice winningPrice) {
-        return Long.valueOf(this.purchasedLottos.stream().filter(lottoPaper ->
-                lottoPaper.getWinningPrice(winningNumbers)
-                        .equals(winningPrice))
-                .count())
-                .intValue() * winningPrice.getPrice();
+    private long getWinningPrice(WinningPrice winningPrice) {
+        return this.purchasedLottos.stream()
+                .filter(lottoPaper ->
+                        lottoPaper.getWinningPrice(winningNumbers)
+                                .equals(winningPrice)
+                )
+                .count() * winningPrice.getPrice();
     }
 }
